% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collabeeration-package.R
\docType{package}
\name{collabeeration-package}
\alias{collabeeration}
\alias{collabeeration-package}
\title{collabeeration: Analyze Beer and Brewery Data}
\description{
This package is designed from the quality control perspective and it aims to provide the tools and resources to analyze brewery and fermentation data.
}
\author{
\strong{Maintainer}: Javiera Rudolph \email{javirudolph@outlook.com}

Authors:
\itemize{
  \item Arthur Rudolph
}

}
\keyword{internal}
